# storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "Your API Title",
        "description": "Your API Description",
        "contact": {
            "name": "Your Name",
            "email": "you@example.com"
        },
        "license": {
            "name": "MIT",
            "url": "https://opensource.org/licenses/MIT"
        },
        "version": "1.0.0"
    },
    "paths": {
...
        "/api/stocks": {
            "get": {
                "tags": [
                    "stocks"
                ],
                "summary": "Get all stocks",
                "description": "Endpoint to retrieve a list of all stocks.",
                "operationId": "bd96270b2b6d7a2fb5564fa499f525c7",
                "responses": {
                    "200": {
                        "description": "List of notifications"
                    }
                }
            },
            "post": {
                "tags": [
                    "stocks"
                ],
                "summary": "Create a new stocks",
                "description": "Endpoint to create a new notifiactions.",
                "operationId": "bcd2af84da9c2ee170ade9d263efffb2",
                "requestBody": {
                    "description": "Request Body Description",
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "properties": {
                                    "product_id": {
                                        "description": "The ID of the product for which the stock is intended",
                                        "type": "integer"
                                    },
                                    "quantity": {
                                        "description": "The quantity of the stock",
                                        "type": "integer"
                                    },
                                    "minimum_quantity": {
                                        "description": "The minimum quantity of the stock",
                                        "type": "integer"
                                    },
                                    "maximum_quantity": {
                                        "description": "The maximum quantity of the stock",
                                        "type": "integer"
                                    },
                                    "created_at": {
                                        "description": "The creation date and time of the stock",
                                        "type": "string",
                                        "format": "date-time"
                                    },
                                    "updated_at": {
                                        "description": "The last update date and time of the stock",
                                        "type": "string",
                                        "format": "date-time"
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Notification created successfully"
                    },
                    "400": {
                        "description": "Bad request, validation failed"
                    }
                }
            }
        },
        "/api/stocks/{stock_id}": {
            "put": {
                "tags": [
                    "stocks"
                ],
                "summary": "Update stock details",
                "description": "Endpoint to update stock details.",
                "operationId": "674e38d11418972982dc2246b1f2a584",
                "parameters": [
                    {
                        "name": "stock_id",
                        "in": "path",
                        "description": "ID of the stock to update",
                        "required": true,
                        "schema": {
                            "type": "integer",
                            "format": "int64"
                        }
                    }
                ],
                "requestBody": {
                    "description": "Request Body Description",
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "properties": {
                                    "product_id": {
                                        "description": "The ID of the product for which the stock is intended",
                                        "type": "integer"
                                    },
                                    "quantity": {
                                        "description": "The quantity of the stock",
                                        "type": "integer"
                                    },
                                    "minimum_quantity": {
                                        "description": "The minimum quantity of the stock",
                                        "type": "integer"
                                    },
                                    "maximum_quantity": {
                                        "description": "The maximum quantity of the stock",
                                        "type": "integer"
                                    },
                                    "created_at": {
                                        "description": "The creation date and time of the stock",
                                        "type": "string",
                                        "format": "date-time"
                                    },
                                    "updated_at": {
                                        "description": "The last update date and time of the stock",
                                        "type": "string",
                                        "format": "date-time"
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Notification details updated successfully"
                    },
                    "400": {
                        "description": "Bad request, validation failed"
                    }
                }
            },
            "delete": {
                "tags": [
                    "stocks"
                ],
                "summary": "Delete a stock",
                "description": "Endpoint to delete a notification by ID.",
                "operationId": "9c61629e71d6064d809d726a8eee9c3b",
                "parameters": [
                    {
                        "name": "stock_id",
                        "in": "path",
                        "description": "ID of the stock to delete",
                        "required": true,
                        "schema": {
                            "type": "integer",
                            "format": "int64"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Notification deleted successfully"
                    },
                    "404": {
                        "description": "Notification not found"
                    }
                }
            }
        },
...
    }
}